## WHAT
Rework the way we do plagiarism benchmarking.
1. Move benchmarking from a rake task into spec files to follow the pattern we've established with ActiveActivitySession controller performance
2. Split benchmarking into segments by plagiarized vs non-plagiarized entries, and by short vs medium vs long entry length, for a total of six segments to benchmark

NOTE: The db structure changes are just updates that were never committed from earlier migrations, so don't need to be examines as part of this PR.
## WHY
1. We want to be consistent with how we implement benchmarking across our system
2. We want to understand how our algorithms perform across known segments in order to better understand what types of improvements are most needed (i.e. do we need to improve the speed of processing long strings)
## HOW
Pulled a bunch of examples from recent Evidence activity sessions and used them to seed the tests.

Current performance results:
```
Average short (50-100 character) plagiarized entries runtime: 0.146 seconds
Average short (50-100 character) non-plagiarized entries runtime: 0.040 seconds
Average medium (101-150 character) plagiarized entries runtime: 0.293 seconds
Average medium (101-150 character) non-plagiarized entries runtime: 0.051 seconds
Average long (151-200 character) plagiarized entries runtime: 0.263 seconds
Average long (151-200 character) non-plagiarized entries runtime: 0.023 seconds
```

### Notion Card Links
https://www.notion.so/quill/Plagiarism-Benchmarking-0069279b9d8b458a8a1be4c660c9469b

PR Checklist | Your Answer
------------ | -------------
Have you added and/or updated tests? |  Yes
Have you deployed to Staging? | No, test and script changes only
Self-Review: Have you done an initial self-review of the code below on Github? | Yes
Spec Review: Have you reviewed the spec and ensured this meets requirements and/or matches design mockups? | N/A